+++ src/main/java/custom/solutions/fileplugin/ConsoleToFileAction.java
@@ -1,15 +1,15 @@
 package custom.solutions.fileplugin;
 
 import com.intellij.execution.impl.ConsoleViewImpl;
-import com.intellij.execution.ui.ConsoleView;
 import com.intellij.icons.AllIcons;
 import com.intellij.openapi.actionSystem.ActionUpdateThread;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.LangDataKeys;
 import com.intellij.openapi.application.ApplicationManager;
+import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.fileChooser.FileChooser;
 import com.intellij.openapi.fileChooser.FileSaverDescriptor;
-import com.intellij.openapi.fileTypes.FileTypes;
+import com.intellij.openapi.fileTypes.PlainTextFileType;
 import com.intellij.openapi.project.DumbAwareAction;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
@@ -19,51 +19,73 @@ import com.intellij.psi.PsiFileFactory;
 import com.intellij.psi.PsiManager;
 import org.jetbrains.annotations.NotNull;
 
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.ResourceBundle;
+
 /**
  * Action.
  */
 public class ConsoleToFileAction extends DumbAwareAction {
+    //        TODO: write a readme.
+    private static final ResourceBundle BUNDLE = ResourceBundle.getBundle("file_plugin");
+    private static final SimpleDateFormat SIMPLE_DATE_FORMAT = new SimpleDateFormat("yyyy.MM.dd_HH.mm");
 
     /**
      * Parent's constructor call.
      */
     public ConsoleToFileAction() {
-        super("Export to File", "Creates file based on UI data", AllIcons.FileTypes.Any_type);
+        super(
+                BUNDLE.getString("plugin.button.title"),
+                BUNDLE.getString("plugin.button.description"),
+                AllIcons.FileTypes.Text
+        );
     }
 
     /**
      * Update representation of button.
      *
-     * @param e {@link AnActionEvent}.
+     * @param event {@link AnActionEvent}.
      */
     @Override
-    public void update(@NotNull AnActionEvent e) {
-        ConsoleView data = e.getData(LangDataKeys.CONSOLE_VIEW);
-        boolean isDataPresented = data != null && data.getContentSize() > 0;
-        e.getPresentation().setEnabled(isDataPresented);
+    public void update(@NotNull AnActionEvent event) {
+        Project project = event.getData(LangDataKeys.PROJECT);
+        ConsoleViewImpl consoleView = getCurrentConsoleView(event);
+        event.getPresentation().setEnabled(project != null && consoleView != null);
     }
 
     /**
      * File created based on console output.
      *
-     * @param e {@link AnActionEvent}.
+     * @param event {@link AnActionEvent}.
      */
     @Override
-    public void actionPerformed(@NotNull AnActionEvent e) {
-        ConsoleView data = e.getData(LangDataKeys.CONSOLE_VIEW);
-        Project project = e.getProject();
+    public void actionPerformed(@NotNull AnActionEvent event) {
+        Project project = event.getData(LangDataKeys.PROJECT);
+//      TODO: add caret model mode
+        ConsoleViewImpl consoleView = getCurrentConsoleView(event);
 
-        FileSaverDescriptor descriptor = new FileSaverDescriptor("Save As", "Save console input");
-        VirtualFile virtualFileDir = FileChooser.chooseFile(descriptor, e.getProject(), null);
+        if (project == null | consoleView == null) {
+            return;
+        }
+
+        FileSaverDescriptor descriptor = new FileSaverDescriptor(
+                BUNDLE.getString("file.saver.descriptor.title"),
+                BUNDLE.getString("file.saver.descriptor.description")
+        );
+        //TODO: think about vf name that we can get from saveFileDialog.
+        VirtualFile virtualFileDir = FileChooser.chooseFile(descriptor, project, null);
+
+        String fileName = createFileName();
+        String initialContent = consoleView.getText();
 
         ApplicationManager.getApplication().runWriteAction(() -> {
-            if (project == null || data == null) {
-                return;
-            }
-            String initialContent = ((ConsoleViewImpl) data).getText();
+            PsiFile fileFromText = PsiFileFactory
+                    .getInstance(project)
+                    .createFileFromText(fileName, PlainTextFileType.INSTANCE, initialContent);
 
-            PsiFile fileFromText = PsiFileFactory.getInstance(project)
-                    .createFileFromText("console_output.txt", FileTypes.PLAIN_TEXT, initialContent);
+//             TODO: Think about this
+//            FileSaverDialog saveFileDialog = FileChooserFactory.getInstance().createSaveFileDialog(descriptor, project);
 
             if (virtualFileDir == null || !virtualFileDir.isDirectory()) {
                 throw new IllegalStateException("Virtual file must be presented as directory.");
@@ -83,6 +105,31 @@ public class ConsoleToFileAction extends DumbAwareAction {
      */
     @Override
     public @NotNull ActionUpdateThread getActionUpdateThread() {
-        return ActionUpdateThread.EDT;
+        /* Out plugin logic may cause an expensive operation. We should not block user interface. */
+        return ActionUpdateThread.BGT;
     }
+
+    /**
+     * Get current project console view.
+     *
+     * @param event {@link AnActionEvent};
+     * @return {@link ConsoleViewImpl}.
+     */
+    private ConsoleViewImpl getCurrentConsoleView(AnActionEvent event) {
+        Editor editor = event.getData(LangDataKeys.EDITOR);
+        if (editor == null) {
+            return null;
+        }
+        return editor.getUserData(ConsoleViewImpl.CONSOLE_VIEW_IN_EDITOR_VIEW);
+    }
+
+    /**
+     * Unique name of file creation.
+     *
+     * @return file name from bundle with current date.
+     */
+    private String createFileName() {
+        return BUNDLE.getString("default.file.name") + SIMPLE_DATE_FORMAT.format(new Date()) + ".txt";
+    }
+
 }