+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 29-09-2024
+### Changed
+- hardcode strings to bundle
+- reorganize console text logic
+- execution of an action thread changed
+- changed file creation pattern
+
 ## [1.0.0] - 28-09-2024
 ### Added
 - file creation based of console output